requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-swagger-ui==4.11.1
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.0
